requirements.txt
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
pyTelegramBotAPI==3.6.7
python-dotenv==0.10.5
requests==2.22.0
service-identity==18.1.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
